OEMx MRC increment 0.0.78.0
[Feature Description]
Update Program-P Platform BIOS with Program-P MRC drop 0.0.78.0

Package/Module: OneSiliconPkg/IpBlock/MemoryInit

[Impacted Platform]
All

Hsd-es-id: 13012955615
Original commit date: Wed Apr 9 14:59:04 2025 +0200
Original commit hash: 901a5facd281c43eae6ee130c7e4c08dd3f1b21b

+++ OneSiliconPkg/IpBlock/MemoryInit/src/ServicesLib/MrcPpr.c
@@ -920,7 +920,7 @@ DispositionFailRangesWithPprFlow (
       NumRepairsSuccessful = 0;
 
       // For each bank pair
-      for (Bank = 0; Bank < (1 << PprAmtData->BaseBits); Bank++) {
+      for (Bank = 0; Bank < (UINT8)(1 << PprAmtData->BaseBits); Bank++) {
         CurAddr.Data = 0;
         CurAddr.Bits.BankPair = Bank;
         CurAddr.Bits.Rank = PprAmtData->Rank;
